ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.5.0
redis==5.0.1
httpx>=0.27.0
pydantic==2.5.2
pydantic-settings==2.1.0
